Bug 1226166 - Move places model from system app to shared, r=daleharvey
