+++ apps/system/test/unit/places_test.js
@@ -1,4 +1,4 @@
-/* global MocksHelper, MockNavigatorDatastore, MockDatastore, BaseModule */
+/* global BaseModule, MocksHelper, MockNavigatorDatastore, MockDatastore */
 /* global asyncStorage, MockService */
 
 'use strict';
@@ -11,6 +11,7 @@ require('/shared/test/unit/mocks/mock_lazy_loader.js');
 
 requireApp('system/js/service.js');
 requireApp('system/js/base_module.js');
+require('/shared/js/places_model.js');
 requireApp('system/js/places.js');
 
 var mocksHelperForPlaces = new MocksHelper([
